refactor: move "frecency_score" to playback
As it is only used there and now there is a more fitting module instead of bloating the normal playlist module.
